backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-spectacular==0.26.5
pymysql==1.1.0
redis==5.0.1
django-redis==5.4.0
Pillow
python-dotenv==1.0.0
